+++ cg/cli/base.py
@@ -24,6 +24,7 @@ from .transfer import transfer
 from .upload import vogue as vogue_command
 from .upload.base import upload
 from .workflow.base import workflow as workflow_cmd
+from .deploy.base import deploy as deploy_cmd
 
 LOG = logging.getLogger(__name__)
 LEVELS = ["DEBUG", "INFO", "WARNING", "ERROR"]
@@ -82,4 +83,6 @@ base.add_command(transfer)
 base.add_command(upload)
 base.add_command(workflow_cmd)
 base.add_command(store_cmd)
+base.add_command(deploy_cmd)
+
 upload.add_command(vogue_command)